change controller for 'formulasi' and 'laktasi'
