+++ composer.json
@@ -17,7 +17,8 @@
         "lavary/laravel-menu": "^1.7",
         "flynsarmy/csv-seeder": "1.0.*",
         "riverskies/laravel-mobile-detect": "^1.3",
-        "unisharp/laravel-ckeditor": "^4.6"
+        "unisharp/laravel-ckeditor": "^4.6",
+        "doctrine/dbal":"2.*"
     },
     "require-dev": {
         "fzaninotto/faker": "~1.4",
