+++ resources/views/home/about.blade.php
