+++ resources/views/home/beranda.blade.php
