+++ resources/views/home/index.blade.php
